Moved contents of AdaptivePooling.cpp to Pooling.cpp +
