+++ lib/Conversion/TorchToLinalg/PopulatePatterns.h
@@ -45,9 +45,6 @@ void populateLinearPatternsAndLegality(TypeConverter &typeConverter,
 void populatePoolingPatternsAndLegality(TypeConverter &typeConverter,
                                         RewritePatternSet &patterns,
                                         ConversionTarget &target);
-void populateAdaptivePoolingPatternsAndLegality(TypeConverter &typeConverter,
-                                                RewritePatternSet &patterns,
-                                                ConversionTarget &target);
 void populateRandomPatternsAndLegality(TypeConverter &typeConverter,
                                        RewritePatternSet &patterns,
                                        ConversionTarget &target);
