+++ lib/Conversion/TorchToLinalg/TorchToLinalg.cpp
@@ -83,9 +83,6 @@ public:
         typeConverter, patterns, target);
     torch_to_linalg::populateTensorConstructorsPatternsAndLegality(
         typeConverter, patterns, target);
-    torch_to_linalg::populateAdaptivePoolingPatternsAndLegality(
-        typeConverter, patterns, target);
-    
 
     if (failed(applyPartialConversion(getOperation(), target,
                                       std::move(patterns))))
